Integrate taxel calibration library and implement BLE-level calibration
- Integrate TaxelCalibrator for converting raw taxel values to Pascal
- Implement calibration at BLE level in SensorDataHandler
- Update PressureSample to include pascalValue field
- Store calibrated Pascal values in WalkModeRepository (not raw values)
- Update CSV generation to use Double (Pascal) instead of Long (raw)
- Add calibration CSV asset file (Sensoles_Sensars_5B_LuT.csv)
- Configure Gradle build files for library integration
- Add core library desugaring for java.time APIs
- Add JUnit test dependency
- Note: sensars-calibration-standalone is a separate git repository (not included)

+++ app/src/main/java/com/sensars/eurostars/data/SessionHistoryManager.kt
@@ -141,7 +141,7 @@ class SessionHistoryManager(private val context: Context) {
         for (i in 1..18) sb.append(",taxel$i")
         sb.append("\n")
         
-        data class Event(val time: Long, val foot: String, val index: Int, val value: Long)
+        data class Event(val time: Long, val foot: String, val index: Int, val value: Double) // Changed to Double for Pascal values
         
         val events = ArrayList<Event>(leftData.size + rightData.size)
         leftData.forEach { events.add(Event(it.timestamp, "Left", it.taxelIndex, it.value)) }
@@ -156,7 +156,7 @@ class SessionHistoryManager(private val context: Context) {
         
         grouped.forEach { (key, groupEvents) ->
             val (time, foot) = key
-            val values = LongArray(18) // Default 0
+            val values = DoubleArray(18) // Default 0.0 for Pascal values
             groupEvents.forEach { event ->
                 if (event.index in 0..17) {
                     values[event.index] = event.value
